requirements.txt
pyvisa
numpy
pywebview
h5py
scipy
pywavelets
